refactor: ast操作の共通関数をlibs/ast-utils.jsに切り出し
best-practice-for-no-unnecessary-variableルールでも使用するため、
スコープ判定やノード探索などの汎用ユーティリティを共通化。
